requirements.txt
numpy==1.21.0
opencv-python==4.5.3.56
mediapipe==0.8.6.2
Pillow==8.2.0
